[feature/linux-kernel-setup] Linux setup script addition (#5)
# Introduction
This PR introduces a way to get a Linux kernel ready to be executed with QEMU. Eventually, the kernel will use shadowfax SBI.

## Description
Using `scripts/setup_linux.sh` a kernel and an initramfs will be built for RISCV architecture using cross-compilation If necessary. The script has the following args:
 - kernel-version (mandatory): kernel version to build; i.e 6.13.2. The version must one from here https://cdn.kernel.org/pub/linux/kernel/
 - busybox-version (optional): busybox version to build. Default 1.36.1
 - cross-compile (optional): cross-compile prefix to build on riscv if non native. Default ""

The script can be used as the following:
```sh
./setup_linux.sh --kernel <kernel-version> [--busybox <busybox>] [--cross-compile <riscv64-toolchain>]
```

For example:
```sh
./scripts/setup_linux.sh --kernel 6.13.2 --cross-compile riscv64-linux-gnu-
```

The output of the script will be a folder called linux-<kernel-version> which will contain the build of the kernel and `initramfs.cpio.gz` archive. A quick QEMU example can be executed:

```sh
qemu-system-riscv64 -M virtual -m 64M \
 -kernel linux-6.13.2/arch/riscv/boot/Image \
 -initrd linux-6.13.2/initramfs.cpio.gz \
 -nographic \
 -append "console=ttyS0 root=/dev/sda earlyprintk=serial net.ifnames=0 nokaslr"

```
Co-authored-by: Granp4sso <stefanomercogliano@gmail.com>